feat(frontend): wallet receive and send shortcuts
